pyvizzu: initial version
