+++ src/ipyvizzu/template.py
@@ -1,25 +1,10 @@
-import enum
+from pyvizzu.template import DisplayTarget as PyvizzuDisplayTarget
+from pyvizzu.template import DisplayTemplate as PyvizzuDisplayTemplate
+from pyvizzu.template import VIZZU as PyvizzuVIZZU
 
 
-class DisplayTarget(str, enum.Enum):
+DisplayTarget = PyvizzuDisplayTarget
 
-    BEGIN = "begin"
-    END = "end"
-    ACTUAL = "actual"
-    MANUAL = "manual"
+DisplayTemplate = PyvizzuDisplayTemplate
 
-
-class DisplayTemplate:
-
-    INIT = """
-        {ipyvizzu_js}
-        window.ipyvizzu = new window.IpyVizzu(element, "{chart_id}", "{vizzu}", "{div_width}", "{div_height}");"""  # pylint: disable=line-too-long
-
-    ANIMATE = "window.ipyvizzu.animate(element, '{chart_id}', '{display_target}', {scroll}, {chart_target}, {chart_anim_opts});"  # pylint: disable=line-too-long
-    STORED = "window.ipyvizzu.stored(element, '{id}')"
-
-    FEATURE = "window.ipyvizzu.feature(element, '{chart_id}', {name}, {enabled});"
-
-    STORE = "window.ipyvizzu.store(element, '{chart_id}', '{id}');"
-
-    CLEAR_INHIBITSCROLL = "window.IpyVizzu.clearInhibitScroll(element);"
+VIZZU = PyvizzuVIZZU
